Bugfix: Respect "Show icon" Enchantrix setting
The code creates a new icon but left it always-visible, even if the
user's "show minimap icon" setting was disabled in Enchantrix. This
meant that after a UI reload, the icon was always visible even if the
user had disabled the minimap icon.

We fix that by always running the visibility-fixing function as soon
as we've taken over Enchantrix's minimap button with our new one.